requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
neo4j==5.14.1

# Supabase
supabase==2.0.3
postgrest==0.13.0
realtime==1.0.0
pgvector==0.2.3

# AI/ML
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
langchain-anthropic==0.1.0
langchain-core==0.1.0
openai==1.3.7
anthropic==0.18.1
tiktoken==0.5.2
ollama==0.1.6

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

